feat(scales): move ticks logic to the scales package
