+++ packages/axes/package.json
@@ -30,13 +30,11 @@
     "@nivo/scales": "0.73.0",
     "@react-spring/web": "9.2.4",
     "d3-format": "^1.4.4",
-    "d3-time": "^1.0.11",
     "d3-time-format": "^3.0.0"
   },
   "devDependencies": {
     "@nivo/core": "0.73.0",
     "@types/d3-format": "^1.4.1",
-    "@types/d3-time": "^1.1.1",
     "@types/d3-time-format": "^2.3.1"
   },
   "peerDependencies": {
