+++ packages/axes/src/types.ts
@@ -1,8 +1,6 @@
 import * as React from 'react'
-import { Scale, ScaleBand, ScalePoint } from '@nivo/scales'
+import { ScaleValue, TicksSpec } from '@nivo/scales'
 import { SpringValues } from '@react-spring/web'
-
-export type AxisValue = string | number | Date
 
 export type GridValuesBuilder<T> = T extends number
     ? number[]
@@ -12,34 +10,18 @@ export type GridValuesBuilder<T> = T extends number
     ? Date[]
     : never
 
-export type GridValues<T extends AxisValue> = number | GridValuesBuilder<T>
+export type GridValues<T extends ScaleValue> = number | GridValuesBuilder<T>
 
 export type Point = {
     x: number
     y: number
 }
 
-export type ScaleWithBandwidth = ScaleBand<any> | ScalePoint<any>
-
-export type AnyScale = Scale<any, any>
-
-export type TicksSpec<Value extends AxisValue> =
-    // exact number of ticks, please note that
-    // depending on the current range of values,
-    // you might not get this exact count
-    | number
-    // string is used for Date based scales,
-    // it can express a time interval,
-    // for example: every 2 weeks
-    | string
-    // override scale ticks with custom explicit values
-    | Value[]
-
 export type AxisLegendPosition = 'start' | 'middle' | 'end'
 
-export type ValueFormatter<Value extends AxisValue> = (value: Value) => Value | string
+export type ValueFormatter<Value extends ScaleValue> = (value: Value) => Value | string
 
-export interface AxisProps<Value extends AxisValue = any> {
+export interface AxisProps<Value extends ScaleValue = any> {
     ticksPosition?: 'before' | 'after'
     tickValues?: TicksSpec<Value>
     tickSize?: number
@@ -53,12 +35,11 @@ export interface AxisProps<Value extends AxisValue = any> {
     ariaHidden?: boolean
 }
 
-export interface CanvasAxisProp<Value extends string | number | Date>
-    extends Omit<AxisProps<Value>, 'legend'> {
+export interface CanvasAxisProp<Value extends ScaleValue> extends Omit<AxisProps<Value>, 'legend'> {
     legend?: string
 }
 
-export interface AxisTickProps<Value extends AxisValue> {
+export interface AxisTickProps<Value extends ScaleValue> {
     tickIndex: number
     value: Value
     format?: ValueFormatter<Value>
