+++ packages/scales/package.json
@@ -28,6 +28,7 @@
   },
   "devDependencies": {
     "@types/d3-scale": "^3.2.2",
+    "@types/d3-time": "^1.1.1",
     "@types/d3-time-format": "^3.0.0"
   },
   "publishConfig": {
