+++ packages/scales/src/index.ts
@@ -6,4 +6,5 @@ export * from './pointScale'
 export * from './timeScale'
 export * from './timeHelpers'
 export * from './bandScale'
+export * from './ticks'
 export * from './types'
