+++ packages/scales/src/types.ts
@@ -121,6 +121,10 @@ export interface ScaleTime<Input> extends D3ScaleTime<Input, number> {
     useUTC: boolean
 }
 
+export type AnyScale = Scale<any, any>
+
+export type ScaleWithBandwidth = ScaleBand<any> | ScalePoint<any>
+
 export type Series<XValue extends ScaleValue, YValue extends ScaleValue> = {
     data: {
         data: {
@@ -144,3 +148,15 @@ export type ComputedSerieAxis<Value extends ScaleValue> = {
     max: Value
     maxStacked?: Value
 }
+
+export type TicksSpec<Value extends ScaleValue> =
+    // exact number of ticks, please note that
+    // depending on the current range of values,
+    // you might not get this exact count
+    | number
+    // string is used for Date based scales,
+    // it can express a time interval,
+    // for example: every 2 weeks
+    | string
+    // override scale ticks with custom explicit values
+    | Value[]
