+++ tsconfig.monorepo.json
@@ -8,11 +8,11 @@
         { "path": "./packages/annotations" },
         { "path": "./packages/scales" },
         { "path": "./packages/axes" },
-        { "path": "./packages/polar-axes" },
         { "path": "./packages/colors" },
         { "path": "./packages/legends" },
         { "path": "./packages/tooltip" },
         { "path": "./packages/arcs" },
+        { "path": "./packages/polar-axes" },
         { "path": "./packages/voronoi" },
 
         // Utility package
